+++ routes/questionset/questionset.route.js
@@ -1,24 +1,21 @@
 var questionset = require('../../models/questionset/questionset.model');
 
-const multer = require('multer');
-
-
-module.exports = function (app) {
+module.exports = function (app,test) {
 
     //pathimage
-    const storage = multer.diskStorage({
-        destination: './public/img/',
-        filename: function (req, file, cb) {
-            cb(null, file.filename + '-' + Date.now() +
-                path.extname(file.originalname));
-        }
-    })
+    // const storage = multer.diskStorage({
+    //     destination: './public/img/',
+    //     filename: function (req, file, cb) {
+    //         cb(null, file.filename + '-' + Date.now() +
+    //             path.extname(file.originalname));
+    //     }
+    // })
 
-    const upload = multer({
-        storage: storage
-    });
-    const test = upload.single('questionset_img');
-    app.use(test);
+    // const upload = multer({
+    //     storage: storage
+    // });
+    // const test = upload.single('questionset_img');
+    // app.use(test);
 
     app.get('/questionset', (req, res) => {
         //res.end(" question sets list");
@@ -39,20 +36,20 @@ module.exports = function (app) {
         })
 
         .post((req, res) => {
-            console.log(req.body);
-            console.log("abc");
+            let filename = '';
+            test(req, res, err => {
+                if (err) {
+                    // res.render('questionsets/add_questionset');
+                } else {
+                    return filename = req.file.filename;                    
+                    // res.render('questionsets/questionset');
+                }
+            })
             questionset.add(req.body)
                 .then(id => {
                     console.log(id);
                     // res.render('questionsets/add_questionset');
-                    test(req, res, err => {
-                        if (err) {
-                            res.render('questionsets/add_questionset');
-                        } else {
-                            console.log(req.file);
-
-                        }
-                    })
+                    
                 })
                 .catch(err => {
                     console.log(err);
